06-anthropic-claude-evaluator/requirements.txt
openai==1.54.0
anthropic==0.40.0
python-dotenv==1.0.1